% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baromax.R
\name{baromax}
\alias{baromax}
\title{Welcome to the baromax function!}
\usage{
baromax(data, x, y, fill, error_width = 0.2)
}
\description{
This function is designed to plot 2x2 barplots with standard error bars using only
one line of code.
}
\details{
To use it, simply specify the dataset, x, y, and fill
These types of plots are designed for cases where y is numeric and
both x and fill are categorical variables.

Change the width of the error bars by specifying error_width = ___
}
