% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histomax.R
\name{histomax}
\alias{histomax}
\title{Welcome to the histomax function!}
\usage{
histomax(data, colorz = "gray85")
}
\description{
This function is designed to plot a histogram for every numeric variable
in a dataset using just one line of code.
The color of the histograms can be set using colorz = ""
}
\details{
Example:
histomax(iris)
}
